+++ hal/src/clocks/pck.rs
@@ -1,18 +1,33 @@
 use super::*;
 use crate::pac::pmc::pmc_pck::CSS_A as PCK_CSS;
 
+use core::ops::RangeInclusive;
+
 impl<I: PckId> Clock for Pck<I> {
     fn freq(&self) -> Hertz {
         self.freq
     }
 }
 
+/// Possible [`Pck`] errors.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum PckError {
+    /// Wanted prescaler is not in the `1..=256` range.
+    PrescalerInvalid,
+}
+
 impl<I: PckId> Token<Pck<I>> {
-    /// Configures PCKx and returns a token. Corresponds to Step 8 in
-    /// 31.17
-    pub fn configure<SRC: PckSource>(self, source: &SRC, pres: u8) -> Pck<I> {
+    /// Configures a [`Pck`]. The input frequency is divided by `pres`.
+    pub fn configure<SRC: PckSource>(self, source: &SRC, pres: u16) -> Result<Pck<I>, PckError> {
+        // C.f. §31.20.13
+        const PCK_VALID_PRESCALER: RangeInclusive<u16> = 1..=256;
+        if !PCK_VALID_PRESCALER.contains(&pres) {
+            return Err(PckError::PrescalerInvalid);
+        }
+        let p: u8 = (pres - 1).try_into().unwrap();
+
         self.pmc().pmc_pck[I::ID as usize].write(|w| unsafe {
-            w.pres().bits(pres);
+            w.pres().bits(p);
             w.css().bits(SRC::PCK_CSS as u8)
         });
 
@@ -23,10 +38,10 @@ impl<I: PckId> Token<Pck<I>> {
 
         self.pmc().pmc_scer.write(|w| unsafe { w.bits(mask) });
         while (self.pmc().pmc_scsr.read().bits() & mask) == 0 {}
-        Pck {
+        Ok(Pck {
             id: PhantomData,
             freq: source.freq() / (pres as u32),
-        }
+        })
     }
 }